Add checks for Text and SearchableText fixes #154
We now check before creating a new field with the above type. If
the analyzers for autosuggest and ngram are not present they are
first created and then the mapping. The app is closed before
applying the mappings and reopend in case of failure or after
the previous operation is resolved